FE-388 FE-390 FE-379 Profile redesign + Token Claiming (#366)
* Profile redesign

* Open Rewards Claiming Web DApp in a WebView

* Fixes on theme parameter

* Change claiming url

* Use feature flag from firebase remote config to enable/disable token claiming.

* Handle redirect URL

* Add CLAIM_APP_URL via an env var to support multiple flavors

* Fixes on redirect url.

* Parse amount on redirect url

* Fix on removing hardcoded return value to show token claiming at all times

* Fix typo in connect wallet of M5

* Fix when user has no stations to show the "buy station" prompt

* Texts for info buttons on total earned & total claimed and fixes on message dialog

* Handle WalletAddressNotFound transparently on UI

+++ app/src/main/java/com/weatherxm/ui/components/RewardsCardView.kt
@@ -85,7 +85,8 @@ open class RewardsCardView : LinearLayout, KoinComponent {
         binding.timestamp.setVisible(data.rewardFormattedTimestamp != null)
 
         val actualReward = data.actualReward ?: 0F
-        binding.reward.text = resources.getString(R.string.reward, formatTokens(actualReward))
+        binding.reward.text =
+            resources.getString(R.string.reward, formatTokens(actualReward.toBigDecimal()))
 
         binding.scoreExplanation.setOnClickListener {
             trackLearnMore(Analytics.ParamValue.REWARDS_SCORE.paramValue)
@@ -139,7 +140,7 @@ open class RewardsCardView : LinearLayout, KoinComponent {
         }
 
         data.periodMaxReward?.let {
-            binding.maxRewards.text = formatTokens(it)
+            binding.maxRewards.text = formatTokens(it.toBigDecimal())
         }
 
         binding.secondaryInfoContainer.setVisible(
